requirements.txt
Flask==0.12.1
tensorflow==1.0.1
tflearn==0.3
Pillow==4.1.0
cloudinary==1.8.0
requests[security]
flask-babel
numpy==1.12.0
